Final fixes for bootstrap
modified:   build.log

1. No idea what this file is for.

modified:   main.c

1. Added external array to hold program as series of bytes. Note (tricky code!!!)
the constant array that contains the code was created by "xxd -i" and the length
of it is greater than the actual size of the array, which means that it will get
filled by zeros. The array length is an even number of flash 2kb sectors, and this
is what we need to transfer.

2. testprog[] needs to be removed not commented out.

3. Added some diags to let console watchers know what is going on.

4. Changed load_resident() to new plan.

Bootstrap now working with new program.